% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferobust.R
\name{im_critical_value}
\alias{im_critical_value}
\title{Imbens-Manski critical value}
\usage{
im_critical_value(delta, se_max, alpha = 0.05)
}
\arguments{
\item{delta}{Width of the identified set.}

\item{se_max}{Maximum SE of the bounds.}

\item{alpha}{Significance level.}
}
\value{
Critical value c_alpha.
}
\description{
Imbens-Manski critical value
}
