% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferobust.R
\name{lambda_from_posterior_sd}
\alias{lambda_from_posterior_sd}
\title{Estimate lambda from posterior standard deviations}
\usage{
lambda_from_posterior_sd(x, posterior_sd, margin = 0.03)
}
\arguments{
\item{x}{Numeric point-estimate vector.}

\item{posterior_sd}{Numeric posterior SD vector.}

\item{margin}{Half-width used to form a reliability range around the point
estimate.}
}
\value{
An object of class \code{ferobust_lambda}.
}
\description{
Estimate lambda from posterior standard deviations
}
