% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferobust.R
\name{lambda_from_replicate}
\alias{lambda_from_replicate}
\title{Estimate lambda from a repeated or proxy measure}
\usage{
lambda_from_replicate(x, replicate, replicate_lambda = NULL, margin = 0.05)
}
\arguments{
\item{x}{Numeric point-estimate vector.}

\item{replicate}{Numeric repeated/proxy measure of the same latent construct.}

\item{replicate_lambda}{Known reliability of the replicate. If omitted, equal
reliability of the two measures is assumed.}

\item{margin}{Half-width used to form a reliability range around the point
estimate.}
}
\value{
An object of class \code{ferobust_lambda}.
}
\description{
Estimate lambda from a repeated or proxy measure
}
