% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferobust.R
\name{print.ferobust_frontier}
\alias{print.ferobust_frontier}
\title{Print a frontier audit}
\usage{
\method{print}{ferobust_frontier}(x, ...)
}
\arguments{
\item{x}{A \code{ferobust_frontier} object.}

\item{...}{Ignored.}
}
\description{
Print a frontier audit
}
