% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferobust.R
\name{print.ferobust_lambda}
\alias{print.ferobust_lambda}
\title{Print a resolved lambda}
\usage{
\method{print}{ferobust_lambda}(x, ...)
}
\arguments{
\item{x}{A \code{ferobust_lambda} object.}

\item{...}{Ignored.}
}
\description{
Print a resolved lambda
}
